requirements.txt
captcha==0.3
dataclasses==0.6
Django==2.1.7
django-tagging==0.4.5
django-videofield==0.1.1
djangorestframework==3.11.0
djongo==1.2.29
get==2019.3.22
null==0.6.1
Pillow==5.4.1
post==2019.3.22
public==2019.3.22
pymongo==3.7.2
pytz==2018.9
query-string==2019.3.22
request==2019.3.22
six==1.12.0
sqlparse==0.2.4
